pyproject.toml
[project]
name = "cncf-landscape-a-to-z"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "devtools>=0.12.2",
    "duckduckgo-search>=8.1.1",
    "fire>=0.7.1",
    "google-genai>=1.61.0",
    "google-generativeai>=0.8.6",
    "jinja2>=3.1.6",
    "logfire>=4.21.0",
    "prefect>=3.6.15",
    "pydantic-ai-slim[web,openai,anthropic,groq,bedrock]>=1.51.0",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "openai>=1.61.0",
    "anthropic>=0.45.0",
    "boto3>=1.36.0",
]

[dependency-groups]
dev = [
    "invoke>=2.2.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
